requirements.txt
flask
python-dotenv
requests
urllib3
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-chroma
langchain-text-splitters
rank_bm25
pydantic>=2.0.0
pypdf
unstructured